requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.4.2
pydantic-settings==2.0.3

# Data Processing
numpy==1.26.0
scikit-learn==1.3.2

# Image Processing
pillow==10.1.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78

# NLP
nltk==3.8.1

# Deep Learning
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2